Local linting check
Before a push happens it's ideal to check and see if it will even
pass a build we'd pay for or that would call the attention of
others on the team.

Developed a pre-push hook (not seen in commit) and updated the remote
build to look a bit more like the final product will.